feat: Add PySQLXJsonEnconder class for JSON encoding
